+++ uizacoresdk/src/main/java/vn/uiza/core/common/Constants.java
@@ -73,7 +73,7 @@ public class Constants {
     public final static String PREFIXS_SHORT = "https:";
 
     public final static String URL_GET_LINK_PLAY_DEV = "https://dev-ucc.uizadev.io/";
-    public final static String URL_GET_LINK_PLAY_STAG = "https://stag-ucc.uiza.io/";
+    public final static String URL_GET_LINK_PLAY_STAG = "https://stag-ucc.uizadev.io/";
     public final static String URL_GET_LINK_PLAY_PROD = "https://ucc.uiza.io/";
 
     public final static String URL_TRACKING_DEV = "https://dev-tracking.uizadev.io/analytic-tracking/";