+++ uizacoresdk/src/main/java/vn/uiza/restapi/uiza/UZService.java
@@ -9,6 +9,7 @@ import retrofit2.http.PUT;
 import retrofit2.http.Query;
 import rx.Observable;
 import vn.uiza.restapi.uiza.model.v3.UizaWorkspaceInfo;
+import vn.uiza.restapi.uiza.model.v3.ad.AdWrapper;
 import vn.uiza.restapi.uiza.model.v3.authentication.gettoken.ResultGetToken;
 import vn.uiza.restapi.uiza.model.v3.linkplay.getlinkplay.ResultGetLinkPlay;
 import vn.uiza.restapi.uiza.model.v3.linkplay.gettokenstreaming.ResultGetTokenStreaming;
@@ -159,4 +160,7 @@ public interface UZService {
 
     @GET("/api/private/v3/player/info/config/")
     Observable<Object> getSkinConfig(@Query("id") String id);
+
+    @GET("/api/private/v3/media/entity/cue-point")
+    Observable<AdWrapper> getCuePoint(@Query("entityId") String entityId);
 }