+++ uizacoresdk/src/main/java/vn/uiza/views/autosize/UZImageButton.java
@@ -3,6 +3,8 @@ package vn.uiza.views.autosize;
 import android.content.Context;
 import android.content.res.Configuration;
 import android.content.res.TypedArray;
+import android.graphics.Color;
+import android.graphics.drawable.Drawable;
 import android.support.v7.widget.AppCompatImageButton;
 import android.util.AttributeSet;
 
@@ -37,16 +39,40 @@ public class UZImageButton extends AppCompatImageButton {
         super(context, attrs, defStyleAttr, defStyleRes);
         initSizeScreenW(attrs);
     }*/
-
+    private Drawable drawableEnabled;
+    private Drawable drawableDisabled;
     private int screenWPortrait;
     private int screenWLandscape;
 
     private boolean isTablet;
     private boolean isUseDefault;
 
+    public void setSrcDrawableEnabled() {
+        if (drawableEnabled != null) {
+            setClickable(true);
+            setFocusable(true);
+            setImageDrawable(drawableEnabled);
+        }
+        clearColorFilter();
+        invalidate();
+    }
+
+    public void setSrcDrawableDisabled() {
+        setClickable(false);
+        setFocusable(false);
+        if (drawableDisabled == null) {
+            setColorFilter(Color.GRAY);
+        } else {
+            setImageDrawable(drawableDisabled);
+            clearColorFilter();
+        }
+        invalidate();
+    }
+
     private void initSizeScreenW(AttributeSet attrs) {
         TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.UZImageButton);
         isUseDefault = a.getBoolean(R.styleable.UZImageButton_useDefaultIB, true);
+        drawableDisabled = a.getDrawable(R.styleable.UZImageButton_srcDisabled);
         //LLog.d(TAG, "initSizeScreenW isUseDefault " + isUseDefault);
         if (!isUseDefault) {
             //LLog.d(TAG, "initSizeScreenW -> return");
@@ -83,6 +109,7 @@ public class UZImageButton extends AppCompatImageButton {
             }
         });
         a.recycle();
+        drawableEnabled = getDrawable();
     }
 
     /*private boolean isFullScreen;
@@ -193,7 +220,7 @@ public class UZImageButton extends AppCompatImageButton {
             int maskedAction = event.getActionMasked();
             if (maskedAction == MotionEvent.ACTION_DOWN) {
                 //setColorTint(ContextCompat.getColor(getContext(), R.color.Gray));
-                this.setBackgroundResource(R.drawable.circle_effect);
+                this.setBackgroundResource(R.drawableEnabled.circle_effect);
             } else if (maskedAction == MotionEvent.ACTION_UP) {
                 //clearColorTint();
                 this.setBackgroundResource(0);
@@ -215,9 +242,29 @@ public class UZImageButton extends AppCompatImageButton {
     protected void onFocusChanged(boolean gainFocus, int direction, @Nullable Rect previouslyFocusedRect) {
         super.onFocusChanged(gainFocus, direction, previouslyFocusedRect);
         if (gainFocus) {
-            this.setBackgroundResource(R.drawable.bt_rate);
+            this.setBackgroundResource(R.drawableEnabled.bt_rate);
         } else {
             this.setBackgroundResource(0);
         }
     }*/
+
+    public void setUIVisible(final boolean isVisible) {
+        setClickable(isVisible);
+        setFocusable(isVisible);
+        if (isVisible) {
+            setSrcDrawableEnabled();
+
+            /*setBackgroundColor(Color.TRANSPARENT);
+            getLayoutParams().width = size;
+            getLayoutParams().height = size;
+            invalidate();*/
+        } else {
+            setImageResource(0);
+
+            /*setBackgroundColor(Color.RED);
+            getLayoutParams().width = 0;
+            getLayoutParams().height = 0;
+            invalidate();*/
+        }
+    }
 }